pyproject.toml
[project]
name = "ocms"
version = "1.0.0"
description = "Online Course Management System API"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.27",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.ruff]
line-length = 120
exclude = [
    ".venv",
    "**/migrations",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]